rename opencl files
