+++ opm/simulators/linalg/bda/opencl/openclCPR.cpp
@@ -30,7 +30,7 @@
 
 #include <opm/simulators/linalg/bda/BdaBridge.hpp>
 #include <opm/simulators/linalg/bda/BlockedMatrix.hpp>
-#include <opm/simulators/linalg/bda/opencl/CPR.hpp>
+#include <opm/simulators/linalg/bda/opencl/openclCPR.hpp>
 #include <opm/simulators/linalg/bda/opencl/OpenclMatrix.hpp>
 #include <opm/simulators/linalg/bda/opencl/openclKernels.hpp>
